# Optimize some R3Element operators and FixedVector scalar products